Move versions under roadmaps in redux
Versions are added under the corresponding roadmap.
