+++ frontend/src/components/TaskValueCreatedVisualization.tsx
@@ -3,10 +3,9 @@ import { shallowEqual, useSelector } from 'react-redux';
 import classNames from 'classnames';
 import Tooltip from '@material-ui/core/Tooltip';
 import { allCustomersSelector } from '../redux/roadmaps/selectors';
-import { Customer } from '../redux/roadmaps/types';
+import { Customer, Version } from '../redux/roadmaps/types';
 import { TaskRatingDimension } from '../../../shared/types/customTypes';
 import { RootState } from '../redux/types';
-import { Version } from '../redux/versions/types';
 import { Dot } from './Dot';
 import { totalValueAndWork } from '../utils/TaskUtils';
 import { percent } from '../utils/string';
