+++ frontend/src/redux/versions/selectors.ts
@@ -1,19 +1,6 @@
 import { chosenRoadmapIdSelector } from '../roadmaps/selectors';
 import { RootState } from '../types';
-import { Version, TimeEstimate } from './types';
-
-export const roadmapsVersionsSelector = (
-  state: RootState,
-): Version[] | undefined => {
-  const chosenRoadmapId = chosenRoadmapIdSelector(state);
-  if (!chosenRoadmapId) {
-    return undefined;
-  }
-  if (!state.versions.versions) return undefined;
-  return state.versions.versions
-    .filter((version) => version.roadmapId === chosenRoadmapId)
-    .sort((a, b) => a.sortingRank - b.sortingRank);
-};
+import { TimeEstimate } from './types';
 
 export const plannerTimeEstimatesSelector = (
   state: RootState,
